Move dashboard related code to components
