+++ src/server/components/dashboard/index.ts
@@ -9,3 +9,4 @@ export {
   listItemsPublishController,
 } from "./dashboard";
 export { dashboardRoutes } from "./routes";
+export { dashboardRouter } from "./router";
