+++ src/server/routes/index.ts
@@ -1,9 +1,9 @@
 import { Express } from "express";
 import healthCheckRouter from "./health-check";
-import { listsRouter } from "../components/lists/router";
 import developmentRouter from "./development";
+import { listsRouter } from "../components/lists/router";
 import { sitemapRouter } from "server/components/sitemap";
-import dashboardRouter from "./dashboard";
+import { dashboardRouter } from "../components/dashboard";
 import { cookiesRouter } from "server/components/cookies";
 
 export const configureRouter = (server: Express): void => {
